+++ docs/website/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.11](https://github.com/ttoss/ttoss/compare/@docs/website@2.2.10...@docs/website@2.2.11) (2026-01-02)
+
+**Note:** Version bump only for package @docs/website
+
+## [2.2.10](https://github.com/ttoss/ttoss/compare/@docs/website@2.2.9...@docs/website@2.2.10) (2026-01-01)
+
+**Note:** Version bump only for package @docs/website
+
+## [2.2.9](https://github.com/ttoss/ttoss/compare/@docs/website@2.2.8...@docs/website@2.2.9) (2026-01-01)
+
+**Note:** Version bump only for package @docs/website
+
+## [2.2.8](https://github.com/ttoss/ttoss/compare/@docs/website@2.2.7...@docs/website@2.2.8) (2025-12-30)
+
+**Note:** Version bump only for package @docs/website
+
+## [2.2.7](https://github.com/ttoss/ttoss/compare/@docs/website@2.2.6...@docs/website@2.2.7) (2025-12-30)
+
+**Note:** Version bump only for package @docs/website
+
+## [2.2.6](https://github.com/ttoss/ttoss/compare/@docs/website@2.2.5...@docs/website@2.2.6) (2025-12-29)
+
+**Note:** Version bump only for package @docs/website
+
+## [2.2.5](https://github.com/ttoss/ttoss/compare/@docs/website@2.2.3...@docs/website@2.2.5) (2025-12-29)
+
+**Note:** Version bump only for package @docs/website
+
+## [2.2.3](https://github.com/ttoss/ttoss/compare/@docs/website@2.2.2...@docs/website@2.2.3) (2025-12-29)
+
+**Note:** Version bump only for package @docs/website
+
 ## [2.2.2](https://github.com/ttoss/ttoss/compare/@docs/website@2.2.1...@docs/website@2.2.2) (2025-12-29)
 
 **Note:** Version bump only for package @docs/website