+++ packages/graphql-api-server/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.10.2 (2026-01-05)
+
+**Note:** Version bump only for package @ttoss/graphql-api-server
+
 ## [0.10.1](https://github.com/ttoss/ttoss/compare/@ttoss/graphql-api-server@0.10.0...@ttoss/graphql-api-server@0.10.1) (2025-12-28)
 
 **Note:** Version bump only for package @ttoss/graphql-api-server