+++ packages/react-auth-strapi/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.58](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-strapi@0.2.57...@ttoss/react-auth-strapi@0.2.58) (2026-01-01)
+
+**Note:** Version bump only for package @ttoss/react-auth-strapi
+
+## [0.2.57](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-strapi@0.2.56...@ttoss/react-auth-strapi@0.2.57) (2025-12-30)
+
+**Note:** Version bump only for package @ttoss/react-auth-strapi
+
+## [0.2.56](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-strapi@0.2.55...@ttoss/react-auth-strapi@0.2.56) (2025-12-29)
+
+**Note:** Version bump only for package @ttoss/react-auth-strapi
+
+## [0.2.55](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-strapi@0.2.54...@ttoss/react-auth-strapi@0.2.55) (2025-12-29)
+
+**Note:** Version bump only for package @ttoss/react-auth-strapi
+
 ## [0.2.54](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth-strapi@0.2.53...@ttoss/react-auth-strapi@0.2.54) (2025-12-23)
 
 **Note:** Version bump only for package @ttoss/react-auth-strapi