+++ packages/react-dashboard/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.18](https://github.com/ttoss/ttoss/compare/@ttoss/react-dashboard@0.1.17...@ttoss/react-dashboard@0.1.18) (2026-01-01)
+
+**Note:** Version bump only for package @ttoss/react-dashboard
+
+## [0.1.17](https://github.com/ttoss/ttoss/compare/@ttoss/react-dashboard@0.1.16...@ttoss/react-dashboard@0.1.17) (2025-12-30)
+
+**Note:** Version bump only for package @ttoss/react-dashboard
+
+## [0.1.16](https://github.com/ttoss/ttoss/compare/@ttoss/react-dashboard@0.1.15...@ttoss/react-dashboard@0.1.16) (2025-12-29)
+
+**Note:** Version bump only for package @ttoss/react-dashboard
+
+## [0.1.15](https://github.com/ttoss/ttoss/compare/@ttoss/react-dashboard@0.1.14...@ttoss/react-dashboard@0.1.15) (2025-12-29)
+
+**Note:** Version bump only for package @ttoss/react-dashboard
+
 ## [0.1.14](https://github.com/ttoss/ttoss/compare/@ttoss/react-dashboard@0.1.13...@ttoss/react-dashboard@0.1.14) (2025-12-23)
 
 **Note:** Version bump only for package @ttoss/react-dashboard