+++ packages/ui/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.4.3](https://github.com/ttoss/ttoss/compare/@ttoss/ui@6.4.1...@ttoss/ui@6.4.3) (2026-01-01)
+
+**Note:** Version bump only for package @ttoss/ui
+
+## 6.4.1 (2025-12-30)
+
+**Note:** Version bump only for package @ttoss/ui
+
 # [6.4.0](https://github.com/ttoss/ttoss/compare/@ttoss/ui@6.3.2...@ttoss/ui@6.4.0) (2025-12-22)
 
 ### Features