+++ terezinha-farm/vite-app/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.100](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.12.99...@terezinha-farm/vite-app@0.12.100) (2026-01-05)
+
+**Note:** Version bump only for package @terezinha-farm/vite-app
+
+## [0.12.99](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.12.98...@terezinha-farm/vite-app@0.12.99) (2026-01-01)
+
+**Note:** Version bump only for package @terezinha-farm/vite-app
+
+## [0.12.98](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.12.97...@terezinha-farm/vite-app@0.12.98) (2025-12-30)
+
+**Note:** Version bump only for package @terezinha-farm/vite-app
+
+## [0.12.97](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.12.96...@terezinha-farm/vite-app@0.12.97) (2025-12-29)
+
+**Note:** Version bump only for package @terezinha-farm/vite-app
+
+## [0.12.96](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.12.95...@terezinha-farm/vite-app@0.12.96) (2025-12-29)
+
+**Note:** Version bump only for package @terezinha-farm/vite-app
+
 ## [0.12.95](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.12.94...@terezinha-farm/vite-app@0.12.95) (2025-12-28)
 
 **Note:** Version bump only for package @terezinha-farm/vite-app